Change custom hooks
