+++ src/pages/MainPage.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
 import { MainFooter } from '../components/mainPageComponents/MainFooter';
-import { useUserData } from '../hooks/useUserData';
+import { useAPIData } from '../hooks/useAPIData';
 import { userDataAPI } from '../api/BootcampAPI';
 
 export default function MainPage() {
-  const { userProfile } = useUserData(userDataAPI);
+  const { Data: userProfile } = useAPIData(userDataAPI);
   return (
     <>
       <Outlet context={{ userProfile: userProfile }}></Outlet>
