+++ src/pages/SharedPage.js
@@ -3,19 +3,21 @@ import SharedHeader from '../components/sharedPageComponents/SharedHeader';
 import SharedContents from '../components/sharedPageComponents/SharedContents';
 import { folderDataAPI } from '../api/BootcampAPI';
 import '../styles/shared.css';
-import { useFolderData } from '../hooks/useFolderData';
+import { useAPIData } from '../hooks/useAPIData';
 import { MainHeader } from '../components/common/MainHeader';
 import { useOutletContext } from 'react-router-dom';
-
 export default function SharedPage() {
-  const { folder, cardData } = useFolderData(folderDataAPI);
+  const { Data: folder } = useAPIData(folderDataAPI);
   const { userProfile } = useOutletContext();
-
   return (
     <>
       <MainHeader userProfile={userProfile} navFixed={true}></MainHeader>
-      <SharedHeader folder={folder}></SharedHeader>
-      <SharedContents links={cardData}></SharedContents>
+      {folder && (
+        <>
+          <SharedHeader folder={folder}></SharedHeader>
+          <SharedContents links={folder.cardData}></SharedContents>
+        </>
+      )}
     </>
   );
 }
